cd87033b0e32525cbe34dd19b389f53d5e772f1c refac: rewrite Collapse widget (edouardparis)
b95c7bd3787bb3d5d4ad7062cd07c5d5fc255fe4 fix: floating scrollbar was missing new spacing method (edouardparis)
61e5827a8feac66f11b548944fa5e95601e671b5 fix: disallow collapse on addr card while editing label (edouardparis)
cb237e59739c96068430c0ba80128f10fd69e93b bump iced v14.0 (edouardparis)

Pull request description:

ACKs for top commit:
  pythcoiner:
    tACK cd87033b

Tree-SHA512: 2a140fca4a4b6dc5c9bf9ed83b4d0f333b0d9e07a14ed48d7801ddb2b4397ad422f211c34ad285baa93533de5d9b3a18a4b65c98f0829377a7e97f0b61205e69

+++ liana-business/business-installer/src/views/mod.rs
@@ -35,10 +35,10 @@ use liana_ui::{
 use uuid::Uuid;
 
 pub const INSTALLER_STEPS: usize = 5;
-pub const MENU_ENTRY_WIDTH: u16 = 600;
-pub const ACCOUNT_ENTRY_WIDTH: u16 = MENU_ENTRY_WIDTH - 80;
-pub const MENU_ENTRY_HEIGHT: u16 = 80;
-const EMAIL_ROW_HEIGHT: u16 = 56;
+pub const MENU_ENTRY_WIDTH: u32 = 600;
+pub const ACCOUNT_ENTRY_WIDTH: u32 = MENU_ENTRY_WIDTH - 80;
+pub const MENU_ENTRY_HEIGHT: u32 = 80;
+const EMAIL_ROW_HEIGHT: u32 = 56;
 
 /// Format last edit information as "Edited by [You|email] [relative_time]".
 /// Returns None if `last_edited` is None.
@@ -64,7 +64,7 @@ pub fn format_last_edit_info(
     })
 }
 
-const EMAIL_HEADER_SPACER: u16 = 30;
+const EMAIL_HEADER_SPACER: u32 = 30;
 
 /// Create a breadcrumb element from path segments.
 /// Renders as "Segment1 > Segment2 > Segment3" with styled separators.
