+++ liana-business/business-installer/src/views/registration/mod.rs
@@ -74,7 +74,7 @@ pub fn registration_view(state: &State) -> Element<'_, Msg> {
     let footer_content = if reg_state.user_devices.is_empty() {
         None
     } else {
-        let spacer = MENU_ENTRY_WIDTH - BtnWidth::XL as u16;
+        let spacer = MENU_ENTRY_WIDTH - BtnWidth::XL as u32;
         let skip_btn = btn_secondary(None, "Skip", BtnWidth::XL, Some(Msg::RegistrationSkipAll));
         let footer = row![
             Space::with_width(Length::Fill),
